+++ vulnerabilities/models.py
@@ -40,6 +40,13 @@ class VulnerabilityQuerySet(DataspacedQuerySet):
             affected_packages_count=Count("affected_packages", distinct=True),
         )
 
+    def order_by_risk(self):
+        return self.order_by(
+            models.F("risk_score").desc(nulls_last=True),
+            models.F("weighted_severity").desc(nulls_last=True),
+            models.F("exploitability").desc(nulls_last=True),
+        )
+
 
 class Vulnerability(HistoryDateFieldsMixin, DataspacedModel):
     """
@@ -94,15 +101,45 @@ class Vulnerability(HistoryDateFieldsMixin, DataspacedModel):
         output_field=models.IntegerField(),
         db_persist=True,
     )
-    min_score = models.FloatField(
+    EXPLOITABILITY_CHOICES = [
+        (0.5, _("No exploits known")),
+        (1.0, _("Potential exploits")),
+        (2.0, _("Known exploits")),
+    ]
+    exploitability = models.DecimalField(
         null=True,
         blank=True,
-        help_text=_("The minimum score of the range."),
+        max_digits=2,
+        decimal_places=1,
+        choices=EXPLOITABILITY_CHOICES,
+        help_text=_(
+            "Exploitability refers to the potential or probability of a software "
+            "package vulnerability being exploited by malicious actors to compromise "
+            "systems, applications, or networks. "
+            "It is determined automatically by discovery of exploits."
+        ),
     )
-    max_score = models.FloatField(
+    weighted_severity = models.DecimalField(
         null=True,
         blank=True,
-        help_text=_("The maximum score of the range."),
+        max_digits=3,
+        decimal_places=1,
+        help_text=_(
+            "Weighted severity is the highest value calculated by multiplying each "
+            "severity by its corresponding weight, divided by 10."
+        ),
+    )
+    risk_score = models.DecimalField(
+        null=True,
+        blank=True,
+        max_digits=3,
+        decimal_places=1,
+        help_text=_(
+            "Risk score from 0.0 to 10.0, with higher values indicating greater "
+            "vulnerability risk. "
+            "This score is the maximum of the weighted severity multiplied by "
+            "exploitability, capped at 10."
+        ),
     )
 
     objects = DataspacedManager.from_queryset(VulnerabilityQuerySet)()
@@ -148,60 +185,14 @@ class Vulnerability(HistoryDateFieldsMixin, DataspacedModel):
         through_defaults = {"dataspace_id": self.dataspace_id}
         self.affected_components.add(*components, through_defaults=through_defaults)
 
-    @staticmethod
-    def range_to_values(self, range_str):
-        try:
-            min_score, max_score = range_str.split("-")
-            return float(min_score.strip()), float(max_score.strip())
-        except Exception:
-            return
-
     @classmethod
     def create_from_data(cls, dataspace, data, validate=False, affecting=None):
-        # Computing the min_score and max_score from the `references` as those data
-        # are not provided by the VulnerableCode API.
-        # https://github.com/aboutcode-org/vulnerablecode/issues/1573
-        # severity_range_score = data.get("severity_range_score")
-        # if severity_range_score:
-        #     min_score, max_score = self.range_to_values(severity_range_score)
-        #     data["min_score"] = min_score
-        #     data["max_score"] = max_score
-
-        severities = [
-            score for reference in data.get("references") for score in reference.get("scores", [])
-        ]
-        if scores := cls.get_severity_scores(severities):
-            data["min_score"] = min(scores)
-            data["max_score"] = max(scores)
-
         instance = super().create_from_data(user=dataspace, data=data, validate=False)
 
         if affecting:
             instance.add_affected(affecting)
 
         return instance
-
-    @staticmethod
-    def get_severity_scores(severities):
-        score_map = {
-            "low": [0.1, 3],
-            "moderate": [4.0, 6.9],
-            "medium": [4.0, 6.9],
-            "high": [7.0, 8.9],
-            "important": [7.0, 8.9],
-            "critical": [9.0, 10.0],
-        }
-
-        consolidated_scores = []
-        for severity in severities:
-            score = severity.get("value")
-            try:
-                consolidated_scores.append(float(score))
-            except ValueError:
-                if score_range := score_map.get(score.lower(), None):
-                    consolidated_scores.extend(score_range)
-
-        return consolidated_scores
 
     def as_cyclonedx(self, affected_instances):
         affects = [
@@ -377,6 +368,17 @@ class AffectedByVulnerabilityMixin(models.Model):
         to="vulnerabilities.Vulnerability",
         related_name="affected_%(class)ss",
         help_text=_("Vulnerabilities affecting this object."),
+    )
+    # Based on vulnerablecode.vulnerabilities.models.Package
+    risk_score = models.DecimalField(
+        null=True,
+        blank=True,
+        max_digits=3,
+        decimal_places=1,
+        help_text=_(
+            "Risk score between 0.0 and 10.0, where higher values "
+            "indicate greater vulnerability risk for the package."
+        ),
     )
 
     class Meta: